renamed the java classpath to daislab.csg, updated dockerfile for rl_manager to include tqdm, rich for the progress bar and fixed reset() observation space to be dtype np.float32 to supress the warning

+++ rl_manager/gym_cloudsimplus/gym_cloudsimplus/envs/threesizeapp.py
@@ -30,7 +30,7 @@ def to_string(java_array):
 
 def to_nparray(raw_obs):
     obs = list(raw_obs)
-    return np.array(obs)
+    return np.array(obs, dtype=np.float32)
 
 
 # Based on https://github.com/openai/gym/blob/master/gym/core.py
